Split files for Accordion and AccordionSection
