+++ packages/react/src/Accordion/Accordion.test.tsx
@@ -1,14 +1,11 @@
-import { fireEvent, render, screen } from '@testing-library/react'
+import { render } from '@testing-library/react'
 import { createRef } from 'react'
 import { Accordion } from './Accordion'
 import '@testing-library/jest-dom'
 
 describe('Accordion', () => {
-  const testLabel = 'Test title'
-  const testContent = 'Test content'
-
   it('renders an accordion', () => {
-    const { container } = render(<Accordion />)
+    const { container } = render(<Accordion headingLevel={1} />)
 
     const accordion = container.querySelector('.amsterdam-accordion')
 
@@ -16,148 +13,21 @@ describe('Accordion', () => {
     expect(accordion).toBeVisible()
   })
 
-  it('renders an accordion section', () => {
-    const { container } = render(
-      <Accordion.Section label={testLabel} headingLevel={1}>
-        {testContent}
-      </Accordion.Section>,
-    )
-
-    const accordionSection = container.querySelector('.amsterdam-accordion__section')
-
-    expect(accordionSection).toBeInTheDocument()
-    expect(accordionSection).toBeVisible()
-  })
-
-  it('renders the accordion section BEM class names', () => {
-    const { container } = render(
-      <Accordion.Section label={testLabel} headingLevel={1}>
-        {testContent}
-      </Accordion.Section>,
-    )
-
-    const heading = screen.getByRole('heading', {
-      name: testLabel,
-    })
-    const button = screen.getByRole('button', {
-      name: testLabel,
-    })
-    const section = container.querySelector('.amsterdam-accordion__panel')
-
-    expect(heading).toHaveClass('amsterdam-accordion__header')
-    expect(button).toHaveClass('amsterdam-accordion__button')
-    expect(section).toBeInTheDocument()
-  })
-
-  it('adds and removes --expanded class when the button is clicked', () => {
-    const { getByText } = render(
-      <Accordion.Section label={testLabel} headingLevel={1}>
-        {testContent}
-      </Accordion.Section>,
-    )
-
-    const button = screen.getByRole('button', {
-      name: testLabel,
-    })
-    const sectionContent = getByText(testContent)
-
-    expect(sectionContent).not.toHaveClass('amsterdam-accordion__panel--expanded')
-    fireEvent.click(button)
-    expect(sectionContent).toHaveClass('amsterdam-accordion__panel--expanded')
-    fireEvent.click(button)
-    expect(sectionContent).not.toHaveClass('amsterdam-accordion__panel--expanded')
-  })
-
-  it('adds --expanded class when expanded prop is true', () => {
-    const { getByText } = render(
-      <Accordion.Section label={testLabel} headingLevel={1} expanded>
-        {testContent}
-      </Accordion.Section>,
-    )
-
-    const sectionContent = getByText(testContent)
-
-    expect(sectionContent).toHaveClass('amsterdam-accordion__panel--expanded')
-  })
-
-  it('renders a section HTML tag by defaultt', () => {
-    render(
-      <Accordion.Section label={testLabel} headingLevel={1}>
-        {testContent}
-      </Accordion.Section>,
-    )
-
-    const sectionQuery = screen.queryByRole('region')
-
-    expect(sectionQuery).toBeInTheDocument()
-  })
-
-  it('doesnt render a section HTML tag when the section prop is false', () => {
-    render(
-      <Accordion.Section label={testLabel} headingLevel={1} section={false}>
-        {testContent}
-      </Accordion.Section>,
-    )
-
-    const sectionQuery = screen.queryByRole('region')
-
-    expect(sectionQuery).not.toBeInTheDocument()
-  })
-
-  it('renders an accessible heading', () => {
-    render(
-      <Accordion.Section label={testLabel} headingLevel={3}>
-        {testContent}
-      </Accordion.Section>,
-    )
-
-    const heading = screen.getByRole('heading', {
-      name: testLabel,
-      level: 3,
-    })
-
-    expect(heading).toBeInTheDocument()
-    expect(heading).toBeVisible()
-  })
-
-  it('renders an icon', () => {
-    render(
-      <Accordion.Section label={testLabel} headingLevel={1}>
-        {testContent}
-      </Accordion.Section>,
-    )
-
-    const button = screen.getByRole('button', {
-      name: testLabel,
-    })
-    const icon = button ? button.querySelector('svg:only-child') : null
-
-    expect(icon).toBeInTheDocument()
-  })
-
   it('can have a additional class name', () => {
-    const { container } = render(
-      <Accordion.Section label={testLabel} headingLevel={1} className="test">
-        {testContent}
-      </Accordion.Section>,
-    )
+    const { container } = render(<Accordion headingLevel={1} className="test" />)
 
-    const accordionSection = container.querySelector('.amsterdam-accordion__section')
+    const accordion = container.querySelector('.amsterdam-accordion')
 
-    expect(accordionSection).toHaveClass('test')
+    expect(accordion).toHaveClass('test')
   })
 
   it('supports ForwardRef in React', () => {
     const ref = createRef<HTMLDivElement>()
 
-    const { container } = render(
-      <Accordion.Section label={testLabel} headingLevel={1} ref={ref}>
-        {testContent}
-      </Accordion.Section>,
-    )
+    const { container } = render(<Accordion headingLevel={1} ref={ref} />)
 
-    const accordionSection = container.querySelector('.amsterdam-accordion__section')
+    const accordion = container.querySelector('.amsterdam-accordion')
 
-    expect(ref.current).toBe(accordionSection)
+    expect(ref.current).toBe(accordion)
   })
 })
