+++ packages/react/src/Accordion/index.ts
@@ -1,2 +1,3 @@
 export { Accordion } from './Accordion'
-export type { AccordionProps, AccordionSectionProps } from './Accordion'
+export type { AccordionProps } from './Accordion'
+export type { AccordionSectionProps } from './AccordionSection'
